+++ web-tier/src/test/java/com/outdoor/foodcalc/service/ProductServiceTest.java
@@ -30,9 +30,9 @@ import static org.mockito.Mockito.*;
 public class ProductServiceTest {
 
     private static final double DELTA = 0.00001;
-    private static final int CATEGORY_1_ID = 12345;
-    private static final int CATEGORY_2_ID = 54321;
-    private static final int CATEGORY_3_ID = 33321;
+    private static final long CATEGORY_1_ID = 12345;
+    private static final long CATEGORY_2_ID = 54321;
+    private static final long CATEGORY_3_ID = 33321;
     private static final String CATEGORY_1_NAME = "First category";
     private static final String CATEGORY_2_NAME = "Second category";
     private static final String CATEGORY_3_NAME = "Empty";
